+++ lib/neo_faker/boolean.ex
@@ -7,6 +7,8 @@ defmodule NeoFaker.Boolean do
   """
   @moduledoc since: "0.5.0"
 
+  alias NeoFaker.Boolean.Utils
+
   @doc """
   Generates a random boolean value.
 
@@ -27,18 +29,6 @@ defmodule NeoFaker.Boolean do
       1
 
   """
-  @spec boolean(pos_integer(), Keyword.t()) :: boolean()
-  def boolean(true_ratio \\ 50, opts \\ [])
-
-  def boolean(true_ratio, []) when true_ratio in 0..100 do
-    :rand.uniform() <= true_ratio / 100
-  end
-
-  def boolean(true_ratio, integer: false) when true_ratio in 0..100, do: boolean()
-
-  def boolean(true_ratio, integer: true) when true_ratio in 0..100 do
-    result = :rand.uniform() <= true_ratio / 100
-
-    if result == true, do: 1, else: 0
-  end
+  @spec boolean(pos_integer(), Keyword.t()) :: boolean() | non_neg_integer()
+  defdelegate boolean(true_ratio \\ 50, opts \\ []), to: Utils, as: :boolean
 end
